Add DebugInformation class
